Tidy up stuff
