+++ src/PlutoSliderServer.jl
@@ -3,9 +3,13 @@ module PlutoSliderServer
 include("./MoreAnalysis.jl")
 import .MoreAnalysis
 include("./FileHelpers.jl")
-import .FileHelpers: find_notebook_files_recursive, list_files_recursive
+import .FileHelpers: find_notebook_files_recursive, list_files_recursive, generate_static_export
+
 include("./Export.jl")
 using .Export
+
+include("./Utils.jl")
+using .Utils
 
 import Pluto
 import Pluto: ServerSession, Firebasey, Token, withtoken, pluto_file_extensions, without_pluto_file_extension
@@ -27,99 +31,16 @@ myhash = base64encode ∘ sha256
 
 ###
 # SESSION DEFINITION
-
-abstract type NotebookSession end
-
-Base.@kwdef struct RunningNotebookSession <: NotebookSession
-    path::String
-    hash::String
-    notebook::Pluto.Notebook
-    original_state
-    token::Token = Token()
-    bond_connections::Dict{Symbol,Vector{Symbol}}
-end
-
-Base.@kwdef struct QueuedNotebookSession <: NotebookSession
-    path::String
-    hash::String
-end
-
-Base.@kwdef struct FinishedNotebookSession <: NotebookSession
-    path::String
-    hash::String
-    original_state
-end
-
-
-
+include("./NotebookSession.jl")
 ###
 # CONFIGURATION
 
-UnionNothingString = Any
-
-@option struct SliderServerSettings
-    exclude::Vector = String[]
-    port::Integer = 2345
-    host = "127.0.0.1"
-    simulated_lag::Real = 0
-    serve_static_export_folder::Bool = true
-end
-
-@option struct ExportSettings
-    output_dir::UnionNothingString = nothing
-    exclude::Vector = String[]
-    ignore_cache::Vector = String[]
-    pluto_cdn_root::UnionNothingString = nothing
-    baked_state::Bool = true
-    offer_binder::Bool = true
-    disable_ui::Bool = true
-    cache_dir::UnionNothingString = nothing
-    slider_server_url::UnionNothingString = nothing
-    binder_url::UnionNothingString = nothing
-    create_index::Bool = true
-end
-
-@option struct PlutoDeploySettings
-    SliderServer::SliderServerSettings = SliderServerSettings()
-    Export::ExportSettings = ExportSettings()
-end
-
-
-function get_configuration(toml_path::Union{Nothing,String}=nothing; kwargs...)
-    if !isnothing(toml_path) && isfile(toml_path)
-        toml_d = TOML.parsefile(toml_path)
-
-        relevant_for_me = filter(toml_d) do (k, v)
-            k ∈ ["SliderServer", "Export"]
-        end
-        relevant_for_pluto = get(toml_d, "Pluto", Dict())
-
-        remaining = setdiff(keys(toml_d), ["SliderServer", "Export", "Pluto"])
-        if !isempty(remaining)
-            @error "Configuration categories not recognised:" remaining
-        end
-
-        kwargs_dict = Configurations.to_dict(Configurations.from_kwargs(PlutoDeploySettings; kwargs...))
-        (
-            Configurations.from_dict(PlutoDeploySettings, merge_recursive(relevant_for_me, kwargs_dict)),
-            Pluto.Configuration.from_flat_kwargs(;(Symbol(k) => v for (k, v) in relevant_for_pluto)...),
-        )
-    else
-        (
-            Configurations.from_kwargs(PlutoDeploySettings; kwargs...),
-            Pluto.Configuration.Options(),
-        )
-    end
-end
-
-merge_recursive(a::AbstractDict, b::AbstractDict) = mergewith(merge_recursive, a, b)
-merge_recursive(a, b) = b
+include("./Settings.jl")
+using .Settings
 
 include("./HTTPRouter.jl")
 
-
-include("./cli.jl")
-export export_directory, run_directory, github_action, cli, FinishedNotebookSession, RunningNotebookSession, QueuedNotebookSession, myhash, MoreAnalysis
+export export_directory, run_directory, github_action, FinishedNotebookSession, RunningNotebookSession, QueuedNotebookSession, myhash, MoreAnalysis
 
 
 """
@@ -281,124 +202,11 @@ function run_directory(
         
         @info "[$(i)/$(length(to_run))] Opening $(path)"
 
-
-        jl_contents = read(joinpath(start_dir, path), String)
-        hash = myhash(jl_contents)
-
-        keep_running = run_server && path ∉ settings.SliderServer.exclude
-        skip_cache = keep_running || path ∈ settings.Export.ignore_cache
-
-        local notebook, original_state
-        
-        cached_state = skip_cache ? nothing : try_fromcache(settings.Export.cache_dir, hash)
-        if cached_state !== nothing
-            @info "Loaded from cache, skipping notebook run" hash
-            original_state = cached_state
-        else
-            try
-                # open and run the notebook (TODO: tell pluto not to write to the notebook file)
-                notebook = Pluto.SessionActions.open(server_session, joinpath(start_dir, path); run_async=false)
-                # get the state object
-                original_state = Pluto.notebook_to_js(notebook)
-                # shut down the notebook
-                if !keep_running
-                    @info "Shutting down notebook process"
-                    Pluto.SessionActions.shutdown(server_session, notebook)
-                end
-
-                try_tocache(settings.Export.cache_dir, hash, original_state)
-            catch e
-                (e isa InterruptException) || rethrow(e)
-                @error "Failed to run notebook!" path exception = (e, catch_backtrace())
-                continue
-            end
-        end
-        
+        add_to_session!(server_session, notebook_sessions, path, settings, pluto_options)
 
         if static_export
-            export_jl_path = let
-                relative_to_notebooks_dir = path
-                joinpath(output_dir, relative_to_notebooks_dir)
-            end
-            export_html_path = let
-                relative_to_notebooks_dir = without_pluto_file_extension(path) * ".html"
-                joinpath(output_dir, relative_to_notebooks_dir)
-            end
-            export_statefile_path = let
-                relative_to_notebooks_dir = without_pluto_file_extension(path) * ".plutostate"
-                joinpath(output_dir, relative_to_notebooks_dir)
-            end
-
-
-            mkpath(dirname(export_jl_path))
-            mkpath(dirname(export_html_path))
-            mkpath(dirname(export_statefile_path))
-
-
-            notebookfile_js = if (settings.Export.offer_binder || settings.Export.slider_server_url !== nothing)
-                repr(basename(export_jl_path))
-            else
-                "undefined"
-            end
-            slider_server_url_js = if settings.Export.slider_server_url !== nothing
-                repr(settings.Export.slider_server_url)
-            else
-                "undefined"
-            end
-            binder_url_js = if settings.Export.offer_binder
-                repr(something(settings.Export.binder_url, "https://mybinder.org/v2/gh/fonsp/pluto-on-binder/v$(string(pluto_version))"))
-            else
-                "undefined"
-            end
-            statefile_js = if !settings.Export.baked_state
-                open(export_statefile_path, "w") do io
-                    Pluto.pack(io, original_state)
-                end
-                repr(basename(export_statefile_path))
-            else
-                statefile64 = base64encode() do io
-                    Pluto.pack(io, original_state)
-                end
-
-                "\"data:;base64,$(statefile64)\""
-            end
-
-            html_contents = generate_html(;
-                pluto_cdn_root=settings.Export.pluto_cdn_root,
-                version=pluto_version,
-                notebookfile_js, statefile_js,
-                slider_server_url_js, binder_url_js,
-                disable_ui=settings.Export.disable_ui
-            )
-            write(export_html_path, html_contents)
-
-            if (settings.Export.offer_binder || settings.Export.slider_server_url !== nothing)
-                write(export_jl_path, jl_contents)
-            end
-
-            @info "Written to $(export_html_path)"
+            generate_static_export(path, settings)
         end
-
-        if keep_running
-            bond_connections = MoreAnalysis.bound_variable_connections_graph(notebook)
-            @info "Bond connections" showall(collect(bond_connections))
-
-            # By setting notebook_sessions[i] to a running session, (modifying the array), the HTTP router will now start serving requests for this notebook.
-            notebook_sessions[i] = RunningNotebookSession(;
-                path,
-                hash,
-                notebook, 
-                original_state, 
-                bond_connections,
-            )
-        else
-            notebook_sessions[i] = FinishedNotebookSession(;
-                path,
-                hash,
-                original_state,
-            )
-        end
-
         @info "[$(i)/$(length(to_run))] Ready $(path)" hash
     end
     @info "-- ALL NOTEBOOKS READY --"
